+++ docutils/test/test_transforms/test_hyperlinks.py
@@ -321,10 +321,7 @@ __ URL
         match in this snippet.
     <system_message level="3" line="5" source="test data" type="ERROR">
         <paragraph>
-            Substitution definition contains illegal element <reference>:
-        <literal_block xml:space="preserve">
-            <reference anonymous="1" name="anonymous">
-                anonymous
+            Anonymous references are not supported in a substitution definition.
         <literal_block xml:space="preserve">
             .. |invalid| replace:: anonymous__
     <paragraph>